lunch_service/requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.9.0
sqlparse==0.5.1
tzdata==2024.1
